Factoring out the gq surface characterization
